Remove FUSE passthrough (CAP_SYS_ADMIN incompatible with AppImage), remove bsplugins, fix bsatk types
- Remove FUSE passthrough code: requesting FUSE_CAP_PASSTHROUGH without
  CAP_SYS_ADMIN caused session cleanup crash on shutdown
- Remove bsplugins library (unneeded port)
- Fix bsatk: use uint32_t instead of unsigned long for BSAULong (LP64 compat)

+++ libs/bsatk/include/bsatk/bsatypes.h
@@ -45,7 +45,7 @@ typedef unsigned long long BSAHash;
 typedef unsigned char BSAUChar;
 typedef unsigned short BSAUShort;
 typedef unsigned int BSAUInt;
-typedef unsigned long BSAULong;
+typedef uint32_t BSAULong;
 typedef unsigned long long BSAHash;
 
 #endif  // WIN32
